Rename NewtonianEuler Solutions directory
