+++ src/PointwiseFunctions/AnalyticSolutions/NewtonianEuler/Solutions.hpp
